Added channel icons to toolbar
Now can create/edit channels by clicking on button vs. using menu.  This is extremely important for new users to get a grasp of layers, and for power users to be more productive.

A feature (that is not complete), is that by right clicking on a channel button, the channel can be disabled/hidden.

Notes:
When right clicking a channel button, the menu option for disable/hide might not be updated
Might add a shift right click option to alpha channel button for 'View Alpha as Overlay'
A view option to enable channel buttons should be added.